requirements.txt
appnope==0.1.3
asttokens==2.4.0
Authlib==1.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
DateTime==5.2
decorator==5.1.1
executing==2.0.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.4
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
WTForms==3.1.0
zope.interface==6.1
